Moved updater state object up so it will get initialized on launch and added SUEnableInstallerLauncherService to Info.plist
