+++ CodeEditModules/Modules/AppPreferences/src/Sections/GeneralPreferences/GeneralPreferencesView.swift
@@ -11,7 +11,7 @@ import Sparkle
 
 /// A view that implements the `General` preference section
 public struct GeneralPreferencesView: View {
-    @StateObject private var updater = ObservableUpdater()
+    @EnvironmentObject var updater: ObservableUpdater
 
     private let inputWidth: Double = 160
     private let textEditorWidth: Double = 220
@@ -334,57 +334,7 @@ private extension GeneralPreferencesView {
         }
     }
 }
-class ObservableUpdater: ObservableObject {
-    private let updater: SPUUpdater
-    @Published var automaticallyChecksForUpdates = false {
-        didSet {
-            updater.automaticallyChecksForUpdates = automaticallyChecksForUpdates
-        }
-    }
-    private var automaticallyChecksForUpdatesObservation: NSKeyValueObservation?
-    @Published var lastUpdateCheckDate: Date?
-    private var lastUpdateCheckDateObservation: NSKeyValueObservation?
-    @Published var includePrereleaseVersions = false {
-        didSet {
-            UserDefaults.standard.setValue(includePrereleaseVersions, forKey: "includePrereleaseVersions")
-            if includePrereleaseVersions {
-                updater.setFeedURL(.prereleaseAppcast)
-            } else {
-                updater.setFeedURL(.appcast)
-            }
-        }
-    }
-    init() {
-        updater = SPUStandardUpdaterController(
-            startingUpdater: true,
-            updaterDelegate: nil,
-            userDriverDelegate: nil
-        ).updater
-        automaticallyChecksForUpdatesObservation = updater.observe(
-            \.automaticallyChecksForUpdates,
-            options: [.initial, .new, .old],
-            changeHandler: { [unowned self] updater, change in
-                guard change.newValue != change.oldValue else { return }
-                self.automaticallyChecksForUpdates = updater.automaticallyChecksForUpdates
-            }
-        )
-        lastUpdateCheckDateObservation = updater.observe(
-            \.lastUpdateCheckDate,
-            options: [.initial, .new, .old],
-            changeHandler: { [unowned self] updater, _ in
-                self.lastUpdateCheckDate = updater.lastUpdateCheckDate
-            }
-        )
-        includePrereleaseVersions = UserDefaults.standard.bool(forKey: "includePrereleaseVersions")
-    }
-    func checkForUpdates() {
-        updater.checkForUpdates()
-    }
-}
-extension URL {
-    static let appcast = URL(string: "https://codeeditapp.github.io/CodeEdit/appcast.xml")!
-    static let prereleaseAppcast = URL(string: "https://codeeditapp.github.io/CodeEdit/appcast_pre.xml")!
-}
+
 func configure<Subject>(_ subject: Subject, configuration: (inout Subject) -> Void) -> Subject {
     var copy = subject
     configuration(&copy)
